backend/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.41
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
pydantic==2.12.1
pydantic-settings==2.13.1
openai==2.3.0
python-dotenv==1.1.0
requests==2.32.3
aiofiles==24.1.0
